requirements.txt
nltk==3.7
numpy==1.23.5
torch==1.13.1
torch-geometric==2.2.0
scikit-learn==1.2.0
torchvision==0.14.0
Pillow==9.3.0
matplotlib==3.6.2
